+++ java/api_gateway/src/main/java/com/solo/coderiver/gateway/ApiGatewayApplication.java
@@ -1,4 +1,4 @@
-package com.solo.coderiver.api_gateway;
+package com.solo.coderiver.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
