+++ java/api_gateway/src/main/java/com/solo/coderiver/gateway/utils/CookieUtils.java
@@ -1,4 +1,4 @@
-package com.solo.coderiver.api_gateway.utils;
+package com.solo.coderiver.gateway.utils;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
